# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Glitch API",
        "description": "The API for connecting to Glitch",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/competitions/{uuid}/uploadMainImage": {
            "post": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "Upload main image to competition.",
                "description": "Upload main image to competition.",
                "operationId": "uploadMainImage",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Image file to upload",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "Image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetitionFull"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The competition does not exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "The competition does not exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied to competition.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Access denied to competition."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/competitions/{uuid}/venues": {
            "get": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "List all the venues associated with a competition.",
                "description": "List all the venues associated with a competition.",
                "operationId": "venueList",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CompetitionVenue"
                                    }
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "No competitions listed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "No competitions listed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "Creating a new venue.",
                "description": "Creating a new venue.",
                "operationId": "createVenue",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompetitionVenue"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetitionVenue"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The competition does not exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Access denined to venue."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "New venue not created"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/competitions/{uuid}/venues/{venue_id}": {
            "get": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "Show a single venue by its ID.",
                "description": "Show a single venue by its ID.",
                "operationId": "showVenue",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "venue_id",
                        "in": "path",
                        "description": "UUID of the venue.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetitionVenue"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "Update the venue.",
                "description": "Update the venue.",
                "operationId": "updateVenue",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "venue_id",
                        "in": "path",
                        "description": "UUID of the venue.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompetitionVenue"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetitionVenue"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The competition does not exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Access denined to venue."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied to competition.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Access denined to venue."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "Deletes the venue from the competition.",
                "description": "Deletes the venue from the competition.",
                "operationId": "removeCompetitionVenue",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "venue_id",
                        "in": "path",
                        "description": "UUID of the venue.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Venue successfully deleted from competition."
                    },
                    "403": {
                        "description": "Access denied to competition.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Access denied to competition."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CompetitionFull": {
                "title": "Competition Extended Model",
                "description": "The competition model with complete information.",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Competition"
                    },
                    {
                        "properties": {
                            "admins": {
                                "description": "A list of users who are following the current user.",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/User"
                                }
                            },
                            "contestants": {
                                "description": "A list of users who the current user is following.",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/User"
                                }
                            },
                            "venues": {
                                "description": "A list of venues associated with the competition.",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            },
                            "rounds": {
                                "description": "A list of rounds associated witht he competition.",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/CompetitionRound"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        }
    }
}